src/main/webapp/app/entities/time-sheet/route/time-sheet-routing-resolve.service.spec.ts
import { TestBed } from '@angular/core/testing';
import { HttpResponse } from '@angular/common/http';
import { HttpClientTestingModule } from '@angular/common/http/testing';
import { ActivatedRouteSnapshot, ActivatedRoute, Router, convertToParamMap } from '@angular/router';
import { RouterTestingModule } from '@angular/router/testing';
import { of } from 'rxjs';

import { ITimeSheet, TimeSheet } from '../time-sheet.model';
import { TimeSheetService } from '../service/time-sheet.service';

import { TimeSheetRoutingResolveService } from './time-sheet-routing-resolve.service';

describe('TimeSheet routing resolve service', () => {
  let mockRouter: Router;
  let mockActivatedRouteSnapshot: ActivatedRouteSnapshot;
  let routingResolveService: TimeSheetRoutingResolveService;
  let service: TimeSheetService;
  let resultTimeSheet: ITimeSheet | undefined;

  beforeEach(() => {
    TestBed.configureTestingModule({
      imports: [HttpClientTestingModule, RouterTestingModule.withRoutes([])],
      providers: [
        {
          provide: ActivatedRoute,
          useValue: {
            snapshot: {
              paramMap: convertToParamMap({}),
            },
          },
        },
      ],
    });
    mockRouter = TestBed.inject(Router);
    jest.spyOn(mockRouter, 'navigate').mockImplementation(() => Promise.resolve(true));
    mockActivatedRouteSnapshot = TestBed.inject(ActivatedRoute).snapshot;
    routingResolveService = TestBed.inject(TimeSheetRoutingResolveService);
    service = TestBed.inject(TimeSheetService);
    resultTimeSheet = undefined;
  });

  describe('resolve', () => {
    it('should return ITimeSheet returned by find', () => {
      // GIVEN
      service.find = jest.fn(id => of(new HttpResponse({ body: { id } })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTimeSheet = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTimeSheet).toEqual({ id: 123 });
    });

    it('should return new ITimeSheet if id is not provided', () => {
      // GIVEN
      service.find = jest.fn();
      mockActivatedRouteSnapshot.params = {};

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTimeSheet = result;
      });

      // THEN
      expect(service.find).not.toBeCalled();
      expect(resultTimeSheet).toEqual(new TimeSheet());
    });

    it('should route to 404 page if data not found in server', () => {
      // GIVEN
      jest.spyOn(service, 'find').mockReturnValue(of(new HttpResponse({ body: null as unknown as TimeSheet })));
      mockActivatedRouteSnapshot.params = { id: 123 };

      // WHEN
      routingResolveService.resolve(mockActivatedRouteSnapshot).subscribe(result => {
        resultTimeSheet = result;
      });

      // THEN
      expect(service.find).toBeCalledWith(123);
      expect(resultTimeSheet).toEqual(undefined);
      expect(mockRouter.navigate).toHaveBeenCalledWith(['404']);
    });
  });
});
